♻️ refactor: user-stock을 asset과 함께 관리하도록 디렉토리 구조 변경
