+++ BE/src/asset/asset.module.ts
@@ -4,11 +4,13 @@ import { AssetController } from './asset.controller';
 import { AssetService } from './asset.service';
 import { AssetRepository } from './asset.repository';
 import { Asset } from './asset.entity';
+import { UserStock } from './user-stock.entity';
+import { UserStockRepository } from './user-stock.repository';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Asset])],
+  imports: [TypeOrmModule.forFeature([Asset, UserStock])],
   controllers: [AssetController],
-  providers: [AssetService, AssetRepository],
-  exports: [AssetRepository],
+  providers: [AssetService, AssetRepository, UserStockRepository],
+  exports: [AssetRepository, UserStockRepository],
 })
 export class AssetModule {}
