+++ BE/src/stock/order/stock-order.module.ts
@@ -7,15 +7,9 @@ import { StockOrderRepository } from './stock-order.repository';
 import { SocketModule } from '../../common/websocket/socket.module';
 import { AssetModule } from '../../asset/asset.module';
 import { StockOrderSocketService } from './stock-order-socket.service';
-import { UserStockModule } from '../../userStock/user-stock.module';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Order]),
-    SocketModule,
-    AssetModule,
-    UserStockModule,
-  ],
+  imports: [TypeOrmModule.forFeature([Order]), SocketModule, AssetModule],
   controllers: [StockOrderController],
   providers: [StockOrderService, StockOrderRepository, StockOrderSocketService],
 })
